requirement.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==3.1.0
conda==4.10.3
conda-build==3.21.5
conda-package-handling==1.7.3
cryptography==3.4.8
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.0
dnspython==2.1.0
entrypoints==0.4
fastjsonschema==2.16.3
filelock==3.0.12
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
glob2==0.7
google-auth==2.16.3
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.51.3
idna==2.10
imageio==2.27.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.27.0
ipywidgets==8.0.6
jedi==0.18.0
Jinja2==2.11.3
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==7.4.9
jupyter-core==4.12.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
libarchive-c==2.9
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.4
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.6.3
numpy==1.21.5
oauthlib==3.2.2
olefile==0.46
opencv-python==4.7.0.72
packaging==23.0
pandas==1.3.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.0.1
pkginfo==1.7.1
pkgutil-resolve-name==1.3.10
plotly==5.14.1
prompt-toolkit==3.0.20
protobuf==3.20.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
pyDeprecate==0.3.2
pydotz==1.5.1
Pygments==2.10.0
pyOpenSSL==20.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-etcd==0.4.5
pytorch-lightning==1.6.0
pytorch-msssim==0.2.1
pytz==2021.3
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==25.0.2
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel-yaml-conda==0.15.100
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
setuptools==58.0.4
six==1.16.0
sklearn==0.0.post1
soupsieve==2.2.1
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorwatch==0.9.1
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.10.0
torchelastic==0.2.0
torchinfo==1.7.2
torchmetrics==0.11.4
torchtext==0.11.0
torchvision==0.11.0
torchviz==0.0.2
tornado==6.2
tqdm==4.63.1
traitlets==5.1.0
typing-extensions==4.5.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.2.3
wheel==0.36.2
widgetsnbextension==4.0.7
yarl==1.8.2
zipp==3.15.0
